+++ packages/kryo-qs/src/test/types/custom.spec.ts
@@ -66,7 +66,7 @@ describe("kryo-qs | Custom", function () {
       write<W>(writer: Writer<W>, value: Complex): W {
         return writer.writeString(value.toString());
       },
-      testError(value: Complex): Error | undefined {
+      testError(value: unknown): Error | undefined {
         if (!(value instanceof Complex)) {
           return createInvalidTypeError("Complex", value);
         }