+++ packages/kryo/src/lib/_helpers/case-style.ts
@@ -24,40 +24,40 @@ export function detectCaseStyle(identifier: string): CaseStyle {
 
 export function split(caseStyle: CaseStyle, identifier: string): string[] {
   switch (caseStyle) {
-  case CaseStyle.ScreamingSnakeCase:
-    return identifier.toLowerCase().split("_");
-  case CaseStyle.SnakeCase:
-    return identifier.toLowerCase().split("_");
-  case CaseStyle.KebabCase:
-    return identifier.toLowerCase().split("-");
-  case CaseStyle.CamelCase:
-    return identifier.split(/(?=[A-Z])/).map((part: string): string => part.toLowerCase());
-  case CaseStyle.PascalCase:
-    return identifier.split(/(?=[A-Z])/).map((part: string): string => part.toLowerCase());
-  default:
-    throw new incident.Incident(`IncompleteSwitch: Received unexpected variant for caseStyle: ${caseStyle}`);
+    case CaseStyle.ScreamingSnakeCase:
+      return identifier.toLowerCase().split("_");
+    case CaseStyle.SnakeCase:
+      return identifier.toLowerCase().split("_");
+    case CaseStyle.KebabCase:
+      return identifier.toLowerCase().split("-");
+    case CaseStyle.CamelCase:
+      return identifier.split(/(?=[A-Z])/).map((part: string): string => part.toLowerCase());
+    case CaseStyle.PascalCase:
+      return identifier.split(/(?=[A-Z])/).map((part: string): string => part.toLowerCase());
+    default:
+      throw new incident.Incident(`IncompleteSwitch: Received unexpected variant for caseStyle: ${caseStyle}`);
   }
 }
 
 export function join(caseStyle: CaseStyle, parts: string[]): string {
   switch (caseStyle) {
-  case CaseStyle.ScreamingSnakeCase:
-    return parts.join("_").toUpperCase();
-  case CaseStyle.SnakeCase:
-    return parts.join("_").toLowerCase();
-  case CaseStyle.KebabCase:
-    return parts.join("-").toLowerCase();
-  case CaseStyle.CamelCase:
-    return parts.map((part: string, index: number): string => {
-      const upperLength: number = index === 0 ? 0 : 1;
-      return part.substr(0, upperLength).toUpperCase() + part.substring(upperLength).toLowerCase();
-    }).join("");
-  case CaseStyle.PascalCase:
-    return parts.map((part: string): string => {
-      return part.substr(0, 1).toUpperCase() + part.substring(1).toLowerCase();
-    }).join("");
-  default:
-    throw new incident.Incident(`IncompleteSwitch: Received unexpected variant for caseStyle: ${caseStyle}`);
+    case CaseStyle.ScreamingSnakeCase:
+      return parts.join("_").toUpperCase();
+    case CaseStyle.SnakeCase:
+      return parts.join("_").toLowerCase();
+    case CaseStyle.KebabCase:
+      return parts.join("-").toLowerCase();
+    case CaseStyle.CamelCase:
+      return parts.map((part: string, index: number): string => {
+        const upperLength: number = index === 0 ? 0 : 1;
+        return part.substr(0, upperLength).toUpperCase() + part.substring(upperLength).toLowerCase();
+      }).join("");
+    case CaseStyle.PascalCase:
+      return parts.map((part: string): string => {
+        return part.substr(0, 1).toUpperCase() + part.substring(1).toLowerCase();
+      }).join("");
+    default:
+      throw new incident.Incident(`IncompleteSwitch: Received unexpected variant for caseStyle: ${caseStyle}`);
   }
 }