+++ packages/kryo/src/lib/custom.ts
@@ -7,14 +7,14 @@ export const name: Name = "custom";
 
 export type Read<T> = <R>(reader: Reader<R>, raw: R) => T;
 export type Write<T> = <W>(writer: Writer<W>, value: T) => W;
-export type TestError<T> = (val: T) => Error | undefined;
+export type TestError = (val: unknown) => Error | undefined;
 export type Equals<T> = (val1: T, val2: T) => boolean;
 export type Clone<T> = (val: T) => T;
 
 export interface CustomTypeOptions<T> {
   read: Read<T>;
   write: Write<T>;
-  testError: TestError<T>;
+  testError: TestError;
   equals: Equals<T>;
   clone: Clone<T>;
 }
@@ -23,7 +23,7 @@ export class CustomType<T> implements Type<T> {
   readonly name: Name = name;
   readonly read!: Read<T>;
   readonly write!: Write<T>;
-  readonly testError!: TestError<T>;
+  readonly testError!: TestError;
   readonly equals!: Equals<T>;
   readonly clone!: Clone<T>;
 
@@ -38,8 +38,8 @@ export class CustomType<T> implements Type<T> {
     }
   }
 
-  test(val: T): boolean {
-    return this.testError(val) === undefined;
+  test(value: unknown): value is T {
+    return this.testError(value) === undefined;
   }
 
   private _applyOptions(): void {